image handler added
